Drop Node.js 6 support
LoopBack4 is dropping support for Node 6 because
it is no longer LTS and doesn't support newer
features. It's best for extensions to align with
it. Also added a dummy unit test, and updated
dependencies.

BREAKING CHANGE: Support for Node.js version lower
than 8.0 has been dropped. Please upgrade to Node
8 or higher.